requirements/development.txt
asgiref==3.7.2
attrs==23.2.0
autopep8==2.0.4
colorlog==6.8.0
Django==5.0.1
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-rest-framework==0.1.0
django-silk==5.0.4
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
gprof2dot==2022.7.29
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
pillow==10.2.0
pycodestyle==2.11.1
PyJWT==2.8.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
